% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acwr.R
\name{acwr}
\alias{acwr}
\title{Acute: Chronic Workload Ratio}
\usage{
acwr(x, n_acute = 7, n_chronic = 28, window_acute = 1,
  window_chronic = 1, fun = "ra", coupling = "coupled")
}
\arguments{
\item{x}{A vector of training load values.}

\item{n_acute}{Number of days included in the acute period (default = 7)}

\item{n_chronic}{Number of days included in the chronic period (default = 28)}

\item{window_acute}{Number of days to average over, for EWMA (default = 1). Must be between 1 and nrow(x), inclusive.
If set to 7, it will wait until 7 values before making a calculation}

\item{window_chronic}{Number of days to average over, for EWMA (default = 1). Must be between 1 and nrow(x), inclusive.
If set to 28, it will wait until 28 values before making a calculation}

\item{fun}{Which aggregation function to use, choose between "ra" for rolling averages
and "ewma" for exponentially weighted moving averages. Defaults to "ra".
If the vector of training loads contains any missing values (NA),
"ra" will be based on values before, and then subsequently values after the missing values,
but will never intermix values before and after missing values in the calculation.
"ewma" handles leading missing values, but any vectors with non-monotonous NA or ending in NA will result in error.}

\item{coupling}{Whether the acute training loads is to be included in the denominator, "coupled" (default)
or not, "uncoupled". Based on equation in https://bjsm.bmj.com/content/53/16/988 .
Assumes uncoupled is the same as the chronic period without the 7 days acute (21 days instead of 28), not
that it skips the first 7 days before including 4 weeks of data (28 days).
Equation is only equivalent when n has reached the full period (e.g. all 28 values are available).}
}
\description{
Function for calculating the acute:chronic workload ratio.
It is desgined based on the most common defintions of ACWR.
The function is not designed to calculate any and every possible method for calculating ACWR.
Based off of Matt Sams R-guide for ACWR: https://mattsams89.netlify.com/post/2018-06-09-r-acwr-intro/r-acwr-intro/
This script uses tidyverse, but for data.table equivalents, see Matt Sams guide.
The user can choose between
Rolling Averages vs. Exponentially weighted moving averages
Coupled (default) vs. uncoupled calulation
number of days in acute period (default 7) and number of days in chronic period (default 28)
For EWMA, can decide whether the calculation should start from the first value, or
not until reaching value n
}
\examples{
tl = sample(500:1300, 28, replace = TRUE)
acwr(tl)
acwr(tl, window_acute = 7, window_chronic = 21, fun = "ewma", coupling = "uncoupled")

}
