% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-datahandling.R
\name{add_event_id}
\alias{add_event_id}
\title{Add Event Id}
\usage{
add_event_id(d, date, event)
}
\arguments{
\item{d}{A dataset with dates and an indicator for the event, class dataframe or tibble. Results will append to this dataset}

\item{date}{Each date of data collection. I.e. date of OSTRC-response.
Must be of class Date. If some dates are missing (e.g. no data collection that day),
these will be added in automatically, and the event status (e.g. injury) will be denoted as missing NA.}

\item{event}{An indicator-variable that states 1 for a event-day, 0 for non-event.
An event is typically an injury or competition/match.
Can be any event that is of interest to the study.}
}
\value{
The input dataset with a new column, which indexes each event interval.
}
\description{
Function to find intervals of unequal lengths for each event in the data.
The function adds a column indexing each interval.
This is useful for finding intervals between multiple injuries or
episodes of pain on the same individual. Or the opposite:
the intervals where the participant is currently injured or in pain.
It can also be used to find the
so-called micro-cycle between matches and competition events. That is,
all days up to first match/competition in one cycle,
all days from first match to second match in the next cycle and so on.
Requires a date-variable and an indicator for the event (injury/match/competition).
}
\examples{
d_matches_sim = tribble(~date, ~match,
                        "2017-08-22", 0,
                        "2017-08-23", 1,
                        "2017-08-24", 0,
                        "2017-08-25", 0,
                        "2017-08-26", 1)
d_matches_sim = d_matches_sim \%>\% mutate(date = lubridate::as_date(date))
add_event_id(d_matches_sim, date, match)
}
