% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match-cycle.R
\name{add_match_cycle}
\alias{add_match_cycle}
\title{Function to add an index to an existing training-load dataset
for each match-cycle. That is, all days up to first match in one cycle,
all days from first match to second match in the next cycle and so on.
Requires a data-variable and an indicator for match}
\usage{
add_match_cycle(d, date, match)
}
\arguments{
\item{d}{A dataset with dates and a match-indicator. Results will append to given dataset.}

\item{date}{Each date training load was collected. Requires to be of class Date. Does not need to
be non-missing}

\item{match}{An indicator-variable that states 1 for a match-day, 0 for non-match.}
}
\description{
Function to add an index to an existing training-load dataset
for each match-cycle. That is, all days up to first match in one cycle,
all days from first match to second match in the next cycle and so on.
Requires a data-variable and an indicator for match
}
