% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro-cycle.R
\name{add_micro_cycle}
\alias{add_micro_cycle}
\title{Add Micro Cycle}
\usage{
add_micro_cycle(d, date, event)
}
\arguments{
\item{d}{A dataset with dates and an indicator for the event. Results will append to given dataset}

\item{date}{Each date training load was collected. Requires to be of class Date. Does not need to
be non-missing}

\item{event}{An indicator-variable that states 1 for a event-day, 0 for non-event.
An event is typically a competition or match.}
}
\description{
Function to add an index describing an interval between two competition or match events,
also known as a match/competition micro-cycle. That is, all days up to first match
in one cycle,all days from first match to second match in the next cycle and so on.
Can also be used to create intervals between injuries or other events.
Requires a date-variable and an indicator for the event (match/competition).
}
