% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{anonymize}
\alias{anonymize}
\title{The helper function anonymize() is time-saving in one-dataset cases.
With this function, there is no need to call the anonymization function twice on the same data.}
\usage{
anonymize(x, startnr = 101)
}
\arguments{
\item{x}{Vector with IDs to anonymize in the future}

\item{startnr}{The Starting number for the anonymized ID. Defaults to 101.}
}
\description{
The helper function anonymize() is time-saving in one-dataset cases.
With this function, there is no need to call the anonymization function twice on the same data.
}
