% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\docType{data}
\name{bjsm_blue}
\alias{bjsm_blue}
\title{British Journal of Sports Medicine (BJSM) blue}
\format{An object of class \code{character} of length 1.}
\usage{
bjsm_blue
}
\description{
British Journal of Sports Medicine (BJSM) blue
}
\keyword{datasets}
