% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\docType{data}
\name{bjsm_green}
\alias{bjsm_green}
\title{British Journal of Sports Medicine (BJSM) green}
\format{An object of class \code{character} of length 1.}
\usage{
bjsm_green
}
\description{
British Journal of Sports Medicine (BJSM) green
}
\keyword{datasets}
