% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-incidence.R
\name{calc_incidence}
\alias{calc_incidence}
\title{Calculate incidence}
\usage{
calc_incidence(d_ostrc, id_participant, time, hp_type)
}
\arguments{
\item{d_ostrc}{a dateframe with OSTRC questionnaire responses}

\item{id_participant}{vector within `d_ostrc` that identifies
a person, athlete, participant, etc.}

\item{time}{vector within `d_ostrc` that identifies a time period,
such as a vector of dates or week-numbers. The incidences will be calculated per
value of this vector, such as per week.}

\item{hp_type}{a binary vector within `d_ostrc` which classifies a type of health problem as 1,
and anything that is not the type of health problem as 0.
This can be health problem (1/0), injury (1/0),
illness (1/0), acute injury (1/0) or any other health problem type that the user wishes
to calculate the incidence on.}
}
\description{
A function to calculate the incidence per given time period, such as per week.
}
\examples{
library(tidyr)
d_ostrc = tribble(~id_participant, ~day_nr, ~hp,
                  1, 1, 1,
                  1, 1, 1,
                  1, 2, 0,
                  1, 3, 1,
                  2, 1, 1,
                  2, 2, 0,
                  2, 3, 1)
calc_incidence(d_ostrc, id_participant, week_nr, hp)
}
