% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-incidence.R
\name{calc_incidence_all}
\alias{calc_incidence_all}
\title{Calculate incidence all}
\usage{
calc_incidence_all(
  d_ostrc,
  id_participant,
  time,
  hp_types,
  group = NULL,
  ci_level = 0.95
)
}
\arguments{
\item{d_ostrc}{a dateframe with OSTRC questionnaire responses}

\item{id_participant}{vector within `d_ostrc` that identifies
a person, athlete, participant, etc.}

\item{time}{vector within `d_ostrc` that identifies a time period,
such as a vector of dates or week-numbers. The incidences will be calculated per
value of this vector, such as per week.
Then, the mean of these incidences will be calculated, resulting in a single number.}

\item{hp_types}{a vector of strings representing variable names of
health problem variables within `d_ostrc`. These variables must classify a
type of health problem as 1,
and anything that is not the type of health problem as 0.
This can be health problem (1/0), injury (1/0),
illness (1/0), acute injury (1/0) or any other health problem type that the user wishes
to calculate the incidence of.}

\item{group}{Optional. A strins representing a variable name of
a variable within `d_ostrc`. The variable must be
class categorical or factor. Examples are "season", "gender" etc.
Output will be calculated per subgroup.}

\item{ci_level}{The level of the confidence intervals. Default is 0.95 for 95% confidence intervals.}
}
\description{
A function to calculate the mean incidence for each health problem type
given in a vector of health problems types. For instance, to provide the
incidence of health problems, substantial health problems, injuries, substantial injuries,
contact vs. noncontact injuries and so on.
Uses `calc_incidence_mean` to calculate for each type.
}
\examples{
library(tidyr)
d_ostrc = tribble(~id_participant, ~week_nr, ~hp, ~hp_sub, ~season,
                 1, 1, 1, 0, 1,
                 1, 2, 1, 1, 1,
                 1, 3, 0, 0, 1,
                 2, 1, 1, 1, 1,
                 2, 2, 1, 1, 1,
                 3, 1, 0, 0, 1,
                 3, 2, 0, 0, 1,
                 1, 1, 1, 0, 2,
                 1, 2, 1, 0, 2,
                 1, 3, 0, 0, 2,
                 2, 1, 1, 0, 2,
                 2, 2, 1, 0, 2,
                 3, 1, 1, 1, 2,
                 3, 2, 1, 1, 2
                 )
hp_types_vector = c("hp", "hp_sub")
calc_incidence_all(d_ostrc, id_participant, week_nr, hp_types_vector)
}
