% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ostrc-questionnaire.R
\name{calc_prevalence}
\alias{calc_prevalence}
\title{Calculate prevalance per time period}
\usage{
calc_prevalence(d_ostrc, id_participant, time, hp_type)
}
\description{
Calculate prevalance per time period
}
