% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-prevalence.R
\name{calc_prevalence_mean}
\alias{calc_prevalence_mean}
\title{Calculate prevalance mean}
\usage{
calc_prevalence_mean(d_ostrc, id_participant, time, hp_type, ci_level = 0.95)
}
\arguments{
\item{d_ostrc}{a dateframe with OSTRC questionnaire responses}

\item{id_participant}{vector within `d_ostrc` that identifies
a person, athlete, participant, etc.}

\item{time}{vector within `d_ostrc` that identifies a time period,
such as a vector of dates or week-numbers. The prevalences will be calculated per
value of this vector, such as per week.
Then, the mean of these prevalences will be calculated, resulting in a single number.}

\item{hp_type}{a binary vector within `d_ostrc` which classifies a type of health problem as 1,
and anything that is not the type of health problem as 0.
This can be health problem (1/0), injury (1/0),
illness (1/0), acute injury (1/0) or any other health problem type that the user wishes
to calculate the prevalance on.}

\item{ci_level}{The level of the confidence intervals. Default is 0.95 for 95% confidence intervals.}
}
\description{
A function to calculate the mean prevalence per given time period, such as per week.
}
