% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ostrc-questionnaire.R
\name{calc_severity_score}
\alias{calc_severity_score}
\title{Calculate severity scores}
\usage{
calc_severity_score(ostrc_1, ostrc_2, ostrc_3, ostrc_4)
}
\arguments{
\item{ostrc_1}{vector with responses to OSTRC questionnaire question 1.}

\item{ostrc_2}{vector with responses to OSTRC questionnaire question 2.}

\item{ostrc_3}{vector with responses to OSTRC questionnaire question 3.}

\item{ostrc_4}{vector with responses to OSTRC questionnaire question 4.}
}
\value{
a vector of severity scores
}
\description{
Calculates severity scores based on
calculation in the original OSTRC questionnaire paper: doi.org/10.1136/bjsports-2012-091524.
}
\examples{
q1 = c(17, 8, 8, 0)
q2 = c(25, 17, 17, 0)
q3 = c(25, 8, 17, 0)
q4 = c(25, 8, 0, 0)

calc_severity_score(q1, q2, q3, q4)
}
