% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{ci_bin}
\alias{ci_bin}
\title{Confidence interval for binomial distribution}
\usage{
ci_bin(x, n)
}
\arguments{
\item{x}{Number of successes/events in the attempt.}

\item{n}{Number of independent attempts.}
}
\description{
Use the Wilson-interval, as recommended in «Binomial confidence intervals
and contingency tests: mathematical fundamentals and the evaluation of
alternative methods», by Sean Wallis, University College London.
}
\details{
Returns a tibble with lower and upper limits for a 95 \%
wilson-confidence interval.
}
