% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure-functions.R
\name{color_darker}
\alias{color_darker}
\title{Darker-color Function}
\usage{
color_darker(colors, level = 5)
}
\arguments{
\item{colors}{a vector of colors.}

\item{level}{how much darker to darken the color. negative values for making the color lighter}
}
\description{
Function for making a color x steps darker or lighter
Uses the CIELAB-colorspace for calculations
(instead of RGB).
}
