% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_basketball}
\alias{d_basketball}
\title{Basketball data}
\format{A data frame with 666 rows and 4 variables:
\describe{
  \item{Season}{The season in which the basketballplayers played}
  \item{training.date}{The date of training}
  \item{athlete}{Which athlete the data belongs to}
  \item{tl}{The training load}
}}
\source{
\url{https://mattsams89.netlify.com/post/2018-06-09-r-acwr-intro/r-acwr-intro/tl-data.csv}
}
\usage{
data(d_basketball)
}
\description{
Undocumented data on the training load of basketball players during a 2017 Fall season.
Available from this webpage: \url{https://mattsams89.netlify.com/post/2018-06-09-r-acwr-intro/r-acwr-intro/}
}
\keyword{datasets}
