% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_study}
\alias{d_study}
\title{TL->Injury data}
\format{A data frame with 83 rows and 53 variables}
\source{
\url{https://mattsams89.netlify.com/post/2018-06-09-r-acwr-intro/r-acwr-intro/tl-data.csv}
}
\usage{
data(d_study)
}
\description{
Data on a review of clinical studies on the relationship between training load or workload and injury.
}
\keyword{datasets}
