% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{ewma}
\alias{ewma}
\title{Exponential-Weighted Moving Averages}
\usage{
ewma(x, n, wilder = FALSE)
}
\arguments{
\item{x}{a vector of values that you wish to calculate EWMA on. For instance, training load values.}

\item{n}{numeric. The size of the window for calculating EWMA. I.e. 28 for 28 days, or 4 for 4 weeks, depending on the chosen time unit.}

\item{wilder}{logical; wilder=FALSE (the default) uses an exponential smoothing ratio of 2/(n+1),
same as in williams et al. 2016. wilder=TRUE uses Welles Wilder's exponential smoothing ratio of 1/n}
}
\description{
Function for calculating exponentially weighted moving averages (EWMA)
on a choosable number of days or weeks.
Warning from the TTR package: "Some indicators (e.g. EMA)
are calculated using the indicators' own previous values, and are therefore
unstable in the short-term. As the indicator receives more data, its output becomes more stable."
}
