% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\docType{data}
\name{expand_bar}
\alias{expand_bar}
\title{Bars should start adjecent to the axis, but have some air next to it.
which means assymetric expand-values.
this is predefined in the object for expand-arguments in ggplot}
\format{An object of class \code{numeric} of length 4.}
\usage{
expand_bar
}
\description{
Bars should start adjecent to the axis, but have some air next to it.
which means assymetric expand-values.
this is predefined in the object for expand-arguments in ggplot
}
\keyword{datasets}
