% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ostrc-questionnaire.R
\name{find_hp}
\alias{find_hp}
\title{Find health problems}
\usage{
find_hp(ostrc_1)
}
\arguments{
\item{ostrc_1}{vector of class numeric with values corresponding to
responses in OSTRC questionnaire, question 1.}
}
\value{
a vector of class numeric where 1 = health problem, 0 = not a health problem.
}
\description{
Function for finding health problems.
Returns a vector with binary values 1 health problem,
0 for not a health problem.
Only responses to the first OSTRC question
"Have you had any difficulties participating in
training and competition due to (location) problems
during the past 7 days?" (version 2.0) or
“Have you had any difficulties participating in
normal training and competition due to injury,
illness or other health problems during the past week?" (version 1.0)
that were not “Full participation without health problems”
is considered a health problem.
}
\examples{
  ostrc_1 = c(0, 8, 17, 25)
  find_hp(ostrc_1)

  ostrc_1_other_values = c(0, 1, 2, 3)
  find_hp(ostrc_1_other_values)
}
