% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ostrc-questionnaire.R
\name{find_hp_substantial}
\alias{find_hp_substantial}
\title{Find substantial health problems}
\usage{
find_hp_substantial(ostrc_1, ostrc_2, ostrc_3, version = "2.0")
}
\arguments{
\item{ostrc_1}{vector of class numeric with responses to
OSTRC-questionnaire question 1: "Have you had any difficulties participating in
training and competition due to (location) problems during the past 7 days?".}

\item{ostrc_2}{vector of class numeric with responses to
OSTRC-questionnaire question 2:
"To what extent have you modified your training or competition
due to (location) problems during the past 7 days?" (version 2.0)
OR “To what extent have you reduced your training volume
due to injury, illness or other health problems during the past week?” (version 1.0).
Standard response values are 0, 8, 17, 25 (version 2.0),
or 0, 13, 17, 19, 25 (version 1.0).
Note that the function assumes that input vectors
with values that are not the standard 0, 8, 17, 25
(or 0, 13, 17, 19, 25 for version 1.0)
are coded with the lowest value corresponding to 0
and highest value corresponding to 25.
An example would be a vector with codes 1, 2, 3, 4
for the 4 potential responses.
The function will throw an error if version 1.0
Q2 is coded with non-standard numeric codes.
These must be coded to 0, 13, 17, 19, 25 manually.}

\item{ostrc_3}{vector of class numeric with responses to
OSTRC-questionnaire question 3:
"To what extent have (location) problems affected
your performance during the past 7 days?" (version 2.0) OR
“To what extent has injury, illness or other health problems
affected your performance the past week?” (version 1.0).
Standard response values are 0, 8, 17, 25 (version 2.0),
or 0, 13, 17, 19, 25 (version 1.0).
Note that the function assumes that input vectors
with values that are not the standard 0, 8, 17, 25
(or 0, 13, 17, 19, 25 for version 1.0)
are coded with the lowest value corresponding to 0
and highest value corresponding to 25.
An example would be a vector with codes 1, 2, 3, 4
for the 4 potential responses.
The function will throw an error if version 1.0
Q3 is coded with non-standard numeric codes.
These must be coded to 0, 13, 17, 19, 25 manually.}

\item{version}{String. Either "2.0" (Default) or "1.0".}
}
\value{
a vector of class numeric with binary codes 1 for
        substantial health problem, 0 for
        non-substantial health problem. Non-health problems are returned as NA.
}
\description{
Function to find substantial health problems given the definition
in the original OSTRC-O paper: doi.org/10.1136/bjsports-2012-091524.
Substantial health problems are defined as injury or illness
that required modified training volumes or performance to a moderate extent or worse,
or to a complete absence from sports.
The function follows the OSTRC version 2.0 recipe of extracting substantial health problems
from the OSTRC questionnaire:
Any reply of "Could not participate due to (location) problems"
on Question 1: "Have you had any difficulties
participating in training and
competition due to (location) problems
during the past 7 days?" OR
a reply of “To a moderate extent” or “To a major extent"
on EITHER Question 2 OR Question 3:
“To what extent have you modified your training or competition
due to (location) problems during the past 7 days?”;
“To what extent has injury, illness or other health problems
affected your performance the past week?”.
In addition, for OSTRC questionnaires version 1.0,
a response of "Cannot participate at all"
on Question 2 and Question 3 is also considered substantial,
regardless of the response on Question 1.
The function is compatible with OSTRC-O and OSTRC-H, version 1.0 and 2.0.
}
\examples{
  ostrc_1 = c(0, 0, 0, 0)
  ostrc_2 = c(0, 0, 0, 25)
  ostrc_3 = c(0, 0, 17, 0)
  find_hp_substantial(ostrc_1, ostrc_2, ostrc_3)

  ostrc_2_v1 = c(0, 0, 0, 19)
  ostrc_3_v1 = c(0, 0, 13, 0)
  find_hp_substantial(ostrc_1, ostrc_2_v1, ostrc_3_v1)

  ostrc_2_othercodes = c(1, 2, 3, 4)
  ostrc_3_othercodes = c(0, 1, 2, 3)
  find_hp_substantial(ostrc_1, ostrc_2_othercodes, ostrc_3_othercodes)

  ostrc_1_missing = c(25, NA, NA, NA)
  ostrc_2_missing = c(NA, NA, NA, 17)
  ostrc_3_missing = c(NA, 8, NA, NA)
  find_hp_substantial(ostrc_1_missing, ostrc_2_missing, ostrc_3_missing)
}
