% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure-functions.R
\name{label_n}
\alias{label_n}
\title{Add n and percentages to labels}
\usage{
label_n(d, x, accuracy = 0.1, perc = TRUE)
}
\arguments{
\item{d}{dataframe with vector of labels. Sample size column must be named "n". Proportion column must be named "prop".
Denominator must be in a column named "denominator".}

\item{x}{vector of labels}

\item{accuracy}{the interval-size for percent-calculations, defaults to 0.1 (1 decimal)}

\item{perc}{whether to add percentages (numerator/ denominator and percentage) or to just add
the sample size for each label (n). Defaults to percentages.}
}
\description{
The function adds the sample size and percentage behind each value in a factor,
purposefully for adding these to each label in a graph. E.g. "Male 5/10 (50%)"
}
\examples{
# data d_study is included in the ostrc-package
library(dplyr)
d_n_sports = d_study \%>\%
             count(sport) \%>\%
             mutate(denominator = sum(n),
                           prop = n/denominator) \%>\%
             arrange(desc(prop))

d_n_sports = d_n_sports \%>\% mutate(labels = label_n(., sport)
d_n_sports
}
