% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{make_anonymize_func}
\alias{make_anonymize_func}
\title{The function make_ano_func() needs two arguments.
The first (x) is a vector with IDs and the other (startnr)
is the starting number for the anonymized IDs (the standard value is 101).
The function provides a anonymization function which will make sure that IDs that exist in more than one dataset
will get the same IDs in each dataset its used on.
developed by Stian Tufte Veisene and Karl Ove Hufthammer}
\usage{
make_anonymize_func(x, startnr = 101)
}
\arguments{
\item{x}{Vector with IDs to anonymize in the future}

\item{startnr}{The Starting number for the anonymized ID. Defaults to 101.}
}
\description{
The function make_ano_func() needs two arguments.
The first (x) is a vector with IDs and the other (startnr)
is the starting number for the anonymized IDs (the standard value is 101).
The function provides a anonymization function which will make sure that IDs that exist in more than one dataset
will get the same IDs in each dataset its used on.
developed by Stian Tufte Veisene and Karl Ove Hufthammer
}
