% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro-cycle.R
\name{micro_cycle_acwr}
\alias{micro_cycle_acwr}
\title{Micro-Cycle Based Acute-Chronic Workload Ratio (ACWR)}
\usage{
micro_cycle_acwr(d_acute_chronic, n_chronic_cycles = 2)
}
\arguments{
\item{n_chronic_cycles}{The number of cycles included in the chronic load - 1,
which is used to determine how many places acute needs to be pushed to
be calculated the correct chronic load. Defaults to 2, which is 3 cycles in the chronic load.}

\item{d}{dataframe with acute and chronic loads as a result of micro_cycle_al and micro_cycle_cl.}
}
\description{
Function to calculate ACWR based on a dataset that includes the acute and chronic loads.
The function assumes an uncoupled ACWR.
}
