% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match-cycle.R
\name{micro_cycle_al}
\alias{micro_cycle_al}
\title{Micro-Cycle Based Acute Loads}
\usage{
micro_cycle_al(d, load, micro_cycle)
}
\arguments{
\item{d}{dataframe with load and micro-cycle columns}

\item{load}{Name of variable in dataset with training load values}

\item{micro_cycle}{Name of variable in dataset with index for the micro-cycle the load value belongs to. This will often be the match-period,
but any index, either of irregular length (e.g. match periods) or not (e.g. weeks) will work.}
}
\description{
Function for calculating the mean load of each given mico cycle. Returns a dataframe with an index for each cycle and the acute load.
}
