% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match-cycle.R
\name{micro_cycle_cl}
\alias{micro_cycle_cl}
\title{Micro-Cycle Based Chronic Loads}
\usage{
micro_cycle_cl(d, load, micro_cycle, n_chronic_cycles = 2)
}
\arguments{
\item{d}{dataframe with load and micro-cycle columns}

\item{load}{Name of variable in dataset with training load values}

\item{micro_cycle}{Name of variable in dataset with index for the micro-cycle the load value belongs to. This will often be the match-period,
but any index, either of irregular length or not (eg. weeks) will work.}

\item{n_chronic_cycles}{The number of cycles, in addition to the first, that is to be included in the chronic time period.
Example, n_chronic_cycles = 2 will calculate EWMA over 1 + 2 = 3 cycles. Defaults to 2. Needs to be atleast 1.}
}
\description{
Function for calculating chronic loads for every 3 match cycles in a sliding window.
The final index does not indicate a single match_cycle, but is rather an index for each chronic window
EWMA is the method of calculations. Missing values are removed from each chronic window before calculation.
}
