% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro-cycle.R
\name{micro_cycle_loads}
\alias{micro_cycle_loads}
\title{Micro-Cycle Based Loads}
\usage{
micro_cycle_loads(d, load, micro_cycle, n_chronic_cycles = 2)
}
\arguments{
\item{d}{dataframe with load and micro-cycle columns}

\item{load}{Name of variable in dataset with training load values}

\item{micro_cycle}{Name of variable in dataset with index for the micro-cycle the load value belongs to. This will often be the match-period,
but any index, either of irregular length or not (eg. weeks) will work.}

\item{n_chronic_cycles}{The number of cycles, in addition to the first, that is to be included in the chronic time period.
Example, n_chronic_cycles = 2 will calculate EWMA over 1 + 2 = 3 cycles. Defaults to 2. Needs to be atleast 1.}
}
\description{
Helper function that combines the power of micro_cycle_al, micro_cycle_cl and micro_cycle_acwr.
Calculates acute and chronic loads based on micro cycles, combines the two and finally calculates ACWR.
}
