% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure-functions.R
\name{move_up}
\alias{move_up}
\title{Move-up Label Function}
\usage{
move_up(y, text, height = 0.015)
}
\arguments{
\item{y}{y-coordinate for the middle of text}

\item{text}{the text inside the graph. Used to count how many lines is needed.}

\item{height}{The height each line takes (in the coordiation system)}
}
\description{
Function that moves labels inside graphs if they collide.
}
