% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\docType{data}
\name{nih_contrast}
\alias{nih_contrast}
\title{NIH colors based on powerpoint stylguide.
2 contrast colors}
\format{An object of class \code{character} of length 2.}
\usage{
nih_contrast
}
\description{
NIH colors based on powerpoint stylguide.
2 contrast colors
}
\keyword{datasets}
