% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\docType{data}
\name{nih_distinct}
\alias{nih_distinct}
\title{NIH colors based on powerpoint stylguide.
6 distinct colors}
\format{An object of class \code{character} of length 6.}
\usage{
nih_distinct
}
\description{
NIH colors based on powerpoint stylguide.
6 distinct colors
}
\keyword{datasets}
