% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\docType{data}
\name{nih_ordinal}
\alias{nih_ordinal}
\title{Ordinal NIH colorscale based on red}
\format{An object of class \code{character} of length 3.}
\usage{
nih_ordinal
}
\description{
Ordinal NIH colorscale based on red
}
\keyword{datasets}
