% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-aggregating.R
\name{ra}
\alias{ra}
\title{Rolling Averages}
\usage{
ra(x, n_days, window = TRUE, ...)
}
\arguments{
\item{x}{a vector of training load values. length(x) can be < n_days}

\item{n_days}{number of days the rolling average will calculate on (e.g. 7 for acute window)}

\item{window}{logical or numeric. If FALSE then FUN is only applied when
 all indexes of the rolling window are within the observed time range.
 If TRUE (default), then the subset of indexes that are in range are passed to FUN.
 A numeric argument to partial can be used to determine the minimal window size
for partial computations. Defaults to TRUE and will calculate RA based on what you have available.}
}
\description{
Function for calculating rolling averages on a chooseable number of days
Based on rollapplyr, not rollmean, as rollmean will only start calculating averages
at n values, while rollapplyr allows the user to decide preliminary values.
}
