% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\docType{data}
\name{remove_x}
\alias{remove_x}
\title{Remove x gridlines from ggplot}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 2.}
\usage{
remove_x
}
\description{
Remove x gridlines from ggplot
}
\keyword{datasets}
