% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\docType{data}
\name{remove_y}
\alias{remove_y}
\title{Remove y gridlines from ggplot}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 2.}
\usage{
remove_y
}
\description{
Remove y gridlines from ggplot
}
\keyword{datasets}
