% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\docType{data}
\name{remove_y_ticks}
\alias{remove_y_ticks}
\title{Remove y tick marks for categories
(equivalent to «major breaks» on the axis).
Useful for bargraphs with categorical values, as
the ticmarks are unecessary}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 1.}
\usage{
remove_y_ticks
}
\description{
Remove y tick marks for categories
(equivalent to «major breaks» on the axis).
Useful for bargraphs with categorical values, as
the ticmarks are unecessary
}
\keyword{datasets}
