% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acwr.R
\name{rolling_diff}
\alias{rolling_diff}
\title{Rolling Difference}
\usage{
rolling_diff(x)
}
\arguments{
\item{x}{a vector of training load values}
}
\description{
Function for calculating the difference between one value in a vector and the previous value.
The purpose is to calculate day-to-day or week-to-week differences in training load
}
