% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{smart_round}
\alias{smart_round}
\title{Smart Rounding Function}
\usage{
smart_round(x)
}
\arguments{
\item{x}{a numeric vector with decimals to be rounded}
}
\description{
Considers the smallest difference between values of a vector when rounding.
Use with caution. Will not work on vectors where differences are smaller than 0.1
}
