% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ostrc-questionnaire.R
\name{standardize_coding}
\alias{standardize_coding}
\title{Standardize coding}
\usage{
standardize_coding(ostrc_q)
}
\arguments{
\item{ostrc_q}{vector of class numeric with responses to any of the 4 OSTRC-questionnaire questions.}
}
\value{
a vector of class numeric with the standard codes of 0, 8, 17 or 25.
}
\description{
Function to standardize codes in OSTRC-questionnaire variables
to 0, 8, 17, and 25 for the corresponding replies of 1, 2, 3, 4 on each question.
Will only work if numeric input vector has exactly 4 unique values.
The function automatically arranges these from smallest to largest and assigns
the smallest value to 0, the next smallest to 8, medium to 17, and highest value to 25.
}
\examples{
ostrc_q = c(0, 1, 2, 3, 2, 2, 3)
standardize_coding(ostrc_q)
}
