% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\name{theme_base}
\alias{theme_base}
\title{Theme Base}
\usage{
theme_base(text_size = 12, panel_spacing = "13")
}
\arguments{
\item{text_size}{The fontsize of text in the figure}

\item{panel_spacing}{The amount of "air" between figures in a panel.
Is given in points, "pt", defaults to "13".
Little spacing may make a cluttered look, but might be necessary for
some figures to fit.}
}
\description{
Function for making a theme for ggplot-plots
Based on standard parameters.
}
