% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_study}
\alias{d_study}
\title{DATA TITLE TO WRITE}
\format{
## `who`
A data frame with 83 rows and 53 columns:
\describe{
  \item{study_id}{TO WRITE}
  \item{author_first}{TO WRITE}
  \item{year}{Year}
  \item{title}{TO WRITE}
  \item{study_design_plan}{TO WRITE}
  \item{study_design_plan_code}{TO WRITE}
  \item{study_design_type}{TO WRITE}
  ...
}
}
\source{
<TO WRITE>
}
\usage{
d_study
}
\description{
DATA DESCRIPTION TO WRITE
}
\keyword{datasets}
