% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acwr.R
\name{ewma}
\alias{ewma}
\title{Exponential-Weighted Moving Averages}
\usage{
ewma(x, n_days, window)
}
\arguments{
\item{x}{a vector of training load values}

\item{n_days}{number of days the rolling average will calculate on (e.g. 7 for acute window)}

\item{window}{Number of periods to average over. Must be between 1 and nrow(x), inclusive.
If set to 7, it will wait until 7 values before making a calculation}
}
\description{
Function for calculating exponentioally weighted moving averages (EWMA) on a choosable number of days
Warning from the TTR package: "Some indicators (e.g. EMA, DEMA, EVWMA, etc.)
are calculated using the indicators' own previous values, and are therefore
unstable in the short-term. As the indicator receives more data, its output becomes more stable."
}
