% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acwr.R
\name{moving_range}
\alias{moving_range}
\title{Rolling Difference}
\usage{
moving_range(x, abs = FALSE)
}
\arguments{
\item{x}{a vector of training load values}

\item{abs}{Whether to calculate the absolute values of the differences TRUE, or not, FALSE. Defaults to FALSE.}
}
\description{
Function for calculating the difference between one value in a vector and the previous value.
The purpose is to calculate day-to-day or week-to-week differences in training load
}
